Mostly working, few rough edges
Confirmed encoding and decoding appears to be working. Register and
deregister are working. GetService and ListServices both work. Need to
smooth out the Namespace* Options to allow just Domain or NamespaceID.
There's an odd 'topic:' registration bug coming up that has a hack in
place, but its not great. Watcher is written but not tested, no
confidence there. Broker-uuid.uuid length bug is also still a gotcha.